Align multiple DNA pattern sequences to one subject

This function is useful plot the alignment position of multiple patterns in one subject. It uses Biostrings::pairwiseAlignment(), retrieves the individual alignment limits and converts that to a ggplot object with a few optional complementing information. It does not work if the patterns have overlapping alignment ranges in the subect!

MultiplePairwiseAlignmentsToOneSubject(
  subject,
  patterns,
  subject.name,
  patterns.names,
  type = "global-local",
  subject.alignment.limits,
  print.pattern.positions = T,
  pattern.positions.size = 2,
  print.subject.min.max = F,
  attach.length.to.name = T,
  tile.border.color = NA,
  title
)

Arguments

subject
a character or DNAStringSet of one subject (to provide a name in final graphics, provide a named DNAStringSet)

patterns
a character vector or DNAStringSet of patterns to align to the subject sequence

subject.name
optional provide a name for the subject; overwrites names provided with subject

patterns.names
optional provide names for the patterns; overwrites names provided with patterns

type
the type of alignment passed to Biostrings::pairwiseAlignment; not every type may work well with this function (if there are overlapping ranges of the alignments to the subject for example)

subject.alignment.limits
numeric vector of nt-positions (lower and upper) of the subject to manually limit (cut) the final graphic

print.pattern.positions
check

pattern.positions.size
check

print.subject.min.max
check

attach.length.to.name
add the length of the string to the name on the axis

tile.border.color
character; tiles from geom_tile are used to plot nts - should they have a border color, e.g. "black"; only useful for short alignment and only an aesthetical thing

title
print a title to the alignment

Value

a list: base.plot ggplot object of alignment shows patterns colored by nt, match.plot ggplot object of alignment shows patterns colored match, mismatch, etc, base.df = df and match.df are the respective data.frames used for plotting, seq min.max.subject.position indicates the outer limits of all aligned patterns (min = start positiion of first aligned pattern, max = end position of the last aligned pattern)

Examples

if (FALSE) {

}